Fix a bug where an interdiff containing two files with identical content but different diff data (such as a file modified, then reverted, then modified again, so different revisions) would cause an error when viewing the interdiff.
This is rare, but will happen when you modify a file on a branch, commit
it, submit it for review, then change that file in some way, commit that,
and then restore it to the first version up for review, and then commit
it and submit it again. The diff data will have changed, but we the
actual chunks won't have changed. So we include it in the changed file
listing, but when we go to fetch the data, we filter it out, since we
don't actually see changes. We then return that the diff doesn't exist,
which causes the diff viewer to show an error.

Since we can't know beforehand that the diff content hasn't changed, we
have to just assume that it has changed. So we no longer filter out
unchanged files, preventing the error. What now happens is that when we
render the template, we see that there are no changed chunks, and we
then remove the diff table and file entry from the diff viewer.

Fixes bug #1032.