Improve the repository form in the administration UI.
This change improves the repository form by separating the fields into
fieldsets, adding help text when appropriate, increasing the width of
the fields, and marking the password field as being a password field
(masking the typed text).

Thanks to Sharan Rao for providing a version of this patch as well! We
seem to have fixed bug #942 at the same time.

Fixes bug #942