Fix the media root when the old root is invalid.
It's possible to break a site by moving it, since the media root is
stored in the siteconfig as an absolute URL. This can break today and
will also break when we restructure the code tree.

This change checks when the old media root is non-existant, and updates
the site based on the new media root, as long as the new media root does
exist.