Patch by Colin Caughie to make parent diffs work with Mercurial.
This introduces a new flag on SCMTools for indicating if diffs should
use changeset IDs for comparing parent diffs. On Git, we compare file
IDs for parent diffs, but on Mercurial, there are only revision IDs and
not file IDs, so we must change the comparison logic.

We now use the ID of the parent diff for all the files we need to
compare, not just the files that are in the parent diff.

A detailed explanation is available in the review request.

Reviewed at http://reviews.review-board.org/r/841/