Remove the last_activity_timestamp query and use last_updated instead.
The last_activity_timestamp query could get really slow on the All
Review Requests page, and really wasn't even necessary. It was a case of
a design evolving several times and then losing sight of the better way
of doing this.

The last_activity_timestamp represented the greater value of the review
request's last_updated field and the latest published review's
timestamp. Really, what we needed to do was just save the review request
when publishing a review and then piggy-back on last_updated.

This change undoes much of last_activity_timestamp, which offers a speed
boost, decreases complexity, and prevents the problem of future code
wanting to influence this value. last_updated really should be the
timestamp representing the last activity on the review request anyway.

This also removes get_last_activity_timestamp() on ReviewRequest, since
it was used in only one place, and was a bit too special-cased to go in
ReviewRequest.

This will introduce an issue in existing installs where suddenly the
Last Updated values in the dashboard will revert back some number of
days, since it will now reflect last_updated and not take into account
existing review timestamps. This is perhaps a little annoying, but will
only be a problem for existing review requests, and the next time
they're updated/reviewed, it will be correct.

Fixes bug #1144.
Reviewed at http://reviews.review-board.org/r/880/