Better define and show Review Board version information.
This defines a more structured VERSION variable and generates both
human-readable versions and normalized packaged versions from it.

The base template has been updated to show this new version, instead of
listing "beta" on the front page. This helps when users want to report
bugs.

This also introduces a /info/ webapi URL that returns the version
information on the server, as well as some basic details about the
server itself (URL and list of administrators).

Reviewed at http://reviews.review-board.org/r/897/