Perform a pre-usage check for patch or patch.exe in the path.
This change introduces a pre-usage check (much like our media directory and
database upgrade checks) that verifies that either patch or patch.exe
(depending on platform) is in the Apache user's path. If not in the path,
some helpful instructions are provided for downloading and installing it.

Reviewed at http://reviews.review-board.org/r/919/