Add some help_text to the path in the Repository admin.
This is a pretty simple change to add some help_text to the Repository form in
the admin UI which explains what the path means. In particular, it points out
that for Git repositories, it should be the path to the .git directory on the
local disk.

This addresses bug 1160.