Fix bug with Mercurial parent diffs when the parent diff includes a binary file
The output of the "hg diff" command is not very helpful when it comes to binary
files; in particular it reports a binary file as having been changed even if it
has actually been added.

Because of this, ReviewBoard will check for the existence of the new file in the
parent revision, and throw an exception because it can't find it.

This patch disables checking for file existence for binary files altogether in
the diffviewer code. I think this makes sense because the diff viewer doesn't
show binary file changes anyway, so shouldn't require them to exist.

Testing done:
- Created a review request against a Mercurial repo in which a binary file has
  been added in the parent diff.
- Also tried a review request against a changeset which itself adds a binary
  file.

Patch from Colin Caughie.
Reviewed at http://reviews.review-board.org/r/920/