Display a pop-up bubble for updates/reviews on review requests.
Today, in order to see that there are updates to a review request, you
have to either keep reloading the review request, reloading the
dashboard, or checking e-mail. These are not very efficient methods.

This change introduces a gmail-style popup bubble that appears when
there are new updates to a review request (and specifically indicates if
there's a new diff), or new reviews/replies. We check every 5 minutes.
The user can choose to ignore the notification or reload the page.

The logic for this is implemented in a reviewrequests/<id>/last-update/
API call that can return the last update to a review request and who
made that update. This can also be useful outside of this change for
third-party integration.

When on the diff viewer page, we only look for updates to diffs. The
idea is that the user is interested at that point in the diff, and would
want to know if they should stop reviewing the current version. Other
changes, such as replies, are not as interesting at that point.

Reviewed at http://reviews.review-board.org/r/900/