Fix the rb-site path in the generated search-cron.conf file.
We had a few issues with the generated search-cron.conf file. The first
line contained a missing quote, and on both lines the path to rb-site would
be invalid. We were using __file__ to get the path to rb-site, which no longer
works now that it's a Python module and not a standalone script. We now
generate the proper path depending on whether it's a .exe or a standard Python
file.

Fixes bug #1181
Reviewed at http://reviews.review-board.org/r/972/